Add a default per request connection on g
Whilst this was previously omitted due to concerns about it being the
limiting factor on the app's concurrency. However, I think it is more
likely to be useful to users who are likely to have apps with all
routes using a connection. It can easily be disabled if it is an
issue as well.