Support SQLite databases
This adds a backend interface structure to the codebase modelled on
that used by Databases. This structure allows sqlite database
connections, via aiosqlite. This should, I hope, make the package more
useful and allows for more backends with time.

I've added a discussion document on why Quart-DB exists rather than
using Databases. In summary Databases has a design decision that makes
usage difficult in practice.

This has meant some compromises as supported SQL syntax is not the
same, with `FOR UPDATE` being the main example. In addition type
conversion is done differently for SQLite whereby the conversion to
SQL is done based on the Python type, rather than the column
type. Finally I've dropped the position parameters support as it is
hard to make syntax, but also as it is error prone.

Note that default users will now need to install quart-db[postgresql].