Bugfix ensure connections are released
Previously the migration errors were expected to result in a failed
startup and hence not matter, and I think the connection was expected
to be handled by the user in the case of an error. Neither of these
really make sense though so it makes sense to ensure the connection is
released on error or otherwise.