Add support for a psycopg backend
This is the other major postgresql DB engine in Python. As I've been
seeing odd behaviour with asyncpg this will be quite helpful in
tracking it down.

At the time of this commit psycopg 3.2 has not been released, and
hence this has been tested against the development branch of psycopg.