Replace fetch_one with fetch_first and fetch_sole
fetch_sole will return either the sole returned row, or raise a
MultipleRowsError (if there are multiple rows returned) or return None
if there are no rows returned.

fetch_first will work as fetch_one does at the moment, but with a
clearer name.

fetch_one will be removed in the next release.